Aggregation Code for Stargazing
Implemented the aggregation algorithm given in the Supernovae paper,
using the JSON format Brian gave (an example is given in the file). So
format might change. Also, reads and writes to a file - will have to be
replaced by http requests in Panoptes (Ed’s idea). Was not sure of the
needed JSON output format so just decided on something simple - can
easily be changed.